refactor dedent function
